refactor: create a file for enums
