+++ src/features/game/fields/Row.ts
@@ -2,10 +2,12 @@ import Component from "../../../shared/Component";
 import { debounceListener } from "../../../shared/helpers";
 import { Observer, Publisher } from "../../../shared/Observer";
 import { div } from "../../../ui/tags";
+
 import WordCard, { type Word } from "../card/WordCard";
+import { RowType } from "../enums";
+
 import HintSettings from "../model/HintSettings";
 import RoundState from "../model/RoundState";
-import { RowType } from "../types";
 
 import styles from "./Row.module.css";
