+++ src/shared/helpers.ts
@@ -1,6 +1,6 @@
 import Component from "./Component";
 import LEVELS from "../../data/levels";
-import { StageStatus } from "../features/game/types";
+import { StageStatus } from "../features/game/enums";
 import { type Word } from "../features/game/card/WordCard";
 import { type Round, type Stage } from "../features/game/model/RoundState";
